fix: dedicated /api/scan-now route for Scan Now button
The Scan Now button was calling POST /api/scrape which requires
CRON_SECRET auth, causing 401 in production. Fix:

- Extract full pipeline logic into lib/scrape-pipeline.ts
- /api/scrape now just checks auth then delegates to the pipeline
  (cron invocations continue to work unchanged)
- New /api/scan-now route calls the same pipeline with no secret
  required — intended for UI-triggered manual scans
- StatsBar now calls /api/scan-now
